util: fixed double evaluation in spdk_min/max
Changed spdk_min/max to evaluate it's arguments exactly once, as
suggested by gcc docs [1].
spdk_min(++a, ++b) wouldn't work as expected prior this patch.

[1] https://gcc.gnu.org/onlinedocs/gcc/Typeof.html

+++ lib/nvme/nvme_ctrlr.c
@@ -1385,7 +1385,8 @@ nvme_ctrlr_init_cap(struct spdk_nvme_ctrlr *ctrlr, const union spdk_nvme_cap_reg
 
 	ctrlr->min_page_size = 1u << (12 + ctrlr->cap.bits.mpsmin);
 
-	ctrlr->opts.io_queue_size = spdk_max(ctrlr->opts.io_queue_size, SPDK_NVME_IO_QUEUE_MIN_ENTRIES);
+	ctrlr->opts.io_queue_size = spdk_max(ctrlr->opts.io_queue_size,
+					     (uint32_t) SPDK_NVME_IO_QUEUE_MIN_ENTRIES);
 	ctrlr->opts.io_queue_size = spdk_min(ctrlr->opts.io_queue_size, ctrlr->cap.bits.mqes + 1u);
 	ctrlr->opts.io_queue_size = spdk_min(ctrlr->opts.io_queue_size, max_io_queue_size);
